fix: move interfaces into separate folder
