+++ src/interfaces/interfaces.tsx
@@ -1,0 +1,18 @@
+export interface Person {
+  birth_year: string;
+  eye_color: string;
+  films: string[];
+  gender: string;
+  hair_color: string;
+  height: string;
+  homeworld: string;
+  mass: string;
+  name: string;
+  skin_color: string;
+  created: string;
+  edited: string;
+  species: string[];
+  starships: string[];
+  url: string;
+  vehicles: string[];
+}
